+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [#671]: https://github.com/stm32-rs/stm32f4xx-hal/pull/671
 
+### Fixed
+- reset timer interrupt in `Counter::start` [#670]
+
+[#670]: https://github.com/stm32-rs/stm32f4xx-hal/pull/670
+
 ## [v0.17.0] - 2023-07-11
 
 ### Changed